requirements.txt
# Multi-Task Facial Emotion Recognition
# Requirements for training and deployment

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# Training Dependencies
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
Pillow>=10.0.0

# Deployment Dependencies
opencv-python>=4.8.0
mediapipe>=0.10.0

# Optional: ONNX Runtime for faster CPU inference
onnxruntime>=1.16.0  # Use onnxruntime-gpu for GPU support
onnx>=1.14.0

# Optional: For ONNX to Keras conversion
# tensorflow>=2.13.0
# onnx2keras>=0.0.24

# Utilities
tqdm>=4.65.0
